First commit
This first commit consists of everything done until now in the initial
version.

Existing: Program currently goes to the links (static) given through a
tex file (urls1.txt). It searches throughout that first page for the
URL inside the iFrame, opens the second URL, and then goes looking for
the calendar at the bottom. It then compares the dates with today’s
(again, static, declared at the top) to check if there is an update.

To do:
1) Change the URLs to be created dynamically with the information from
a text file given by the user.

2) Create the necessary text files that contain the information needed
to create the URLs. These must be gotten by reading the source html
from the website. Look out for the codes for each option.

3) Some GUI options:
— 3.1) Date selector
— 3.2) File selector
— 3.3) Option to use the raw data to create URLs and then scan them, or
to use a file with the URLs already created (to reduce subsequent
searches’ waiting time).
— 3.4) Might be cool to add the actual updates that happened (if there
were any). These are on the same table, on another column. Check
website for a graphical representation.

—————Notes—————
URLs and Dates are currently static.

The Date class was created by me (Debml), might be useful to use Java’s
default Date class (might reduce errors?)

Must add the Jsoup library for this code to work.

Important links:
http://www.dgepj.cjf.gob.mx/internet/expedientes/exp_ini.asp?Exp=1

http://jsoup.org/